+++ adslib/build.gradle
@@ -1,0 +1,43 @@
+apply plugin: 'com.android.library'
+
+android {
+    compileSdkVersion 24
+    buildToolsVersion "25.0.0"
+
+    defaultConfig {
+//        applicationId "cn.evolvingera.libchinaads"
+        minSdkVersion 18
+        targetSdkVersion 24
+        versionCode 1
+        versionName "1.0.0"
+        multiDexEnabled true //  multi-dex
+
+    }
+//    buildTypes {
+//        release {
+//            minifyEnabled false
+//            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+//        }
+//    }
+    android {
+        lintOptions {
+            abortOnError false
+        }
+    }
+
+
+}
+
+dependencies {
+//    compile fileTree(dir: 'libs', include: ['*.jar'])
+//    compile 'com.android.support:appcompat-v7:24.0.0'
+//    compile "com.android.support:recyclerview-v7:24.0.0"
+
+//    compile 'com.google.android.gms:play-services:10.2.6'
+    compile 'com.android.support:multidex:1.0.1'
+    compile "com.squareup.picasso:picasso:2.5.2"
+    compile files('libs/InMobi-Android-6.8.0-B1-China.jar')
+    compile 'com.google.android.gms:play-services:8.4.0'
+
+//    compile "com.inmobi.monetization:inmobi-ads:6.1.1"
+}
