+++ app/src/main/java/net/droidman/testjitpack/App.java
@@ -1,0 +1,17 @@
+package net.droidman.testjitpack;
+
+import android.app.Application;
+
+import cn.evolvingera.libads.adscore.AdsManager;
+
+/**
+ * Created by ouyangwenyuan on 2017/6/13.
+ */
+
+public class App extends Application {
+    //ca-app-pub-3618984151272025~2700887995
+    public void onCreate() {
+        super.onCreate();
+
+    }
+}
